requirements.txt
numpy
streamlit
openai
faiss-cpu